implementando soft delete na App e delete offline(quando deleta na App tbm deleta no servidor quando a App voltar a ter internet)

+++ app/src/main/java/br/com/agenda/dao/AlunoDAO.java
@@ -16,7 +16,7 @@ import br.com.agenda.modelo.Aluno;
 
 public class AlunoDAO extends SQLiteOpenHelper{
     public AlunoDAO(Context context) {
-        super(context, "Agenda", null, 5);
+        super(context, "Agenda", null, 6);
     }
 
     @Override
@@ -28,7 +28,8 @@ public class AlunoDAO extends SQLiteOpenHelper{
                 "site TEXT, " +
                 "caminhoFoto TEXT, " +
                 "nota REAL," +
-                "sincronizado INT DEFAULT 0);";
+                "sincronizado INT DEFAULT 0," +
+                "desativado INT DEFAULT 0);";
         db.execSQL(sql);
     }
 
@@ -77,6 +78,9 @@ public class AlunoDAO extends SQLiteOpenHelper{
             case 4:
                 String adicionaCampoSincronizado = "ALTER TABLE Alunos ADD COLUMN sincronizado INT DEFAULT 0";
                 db.execSQL(adicionaCampoSincronizado);
+            case 5:
+                String adicionaCampoDesativado = "ALTER TABLE Alunos ADD COLUMN desativado INT DEFAULT 0";
+                db.execSQL(adicionaCampoDesativado);
 
         }
     }
@@ -93,7 +97,12 @@ public class AlunoDAO extends SQLiteOpenHelper{
         SQLiteDatabase db = getWritableDatabase();
         String[] params = {aluno.getId().toString()};
 
-        db.delete("Alunos", "id = ?", params);
+        if(aluno.estaDesativado())
+            db.delete("Alunos", "id = ?", params);
+        else {
+            aluno.desativa();
+            altera(aluno);
+        }
     }
 
     public void altera(Aluno aluno){
@@ -107,7 +116,7 @@ public class AlunoDAO extends SQLiteOpenHelper{
 
     public List<Aluno> buscaAlunos(){
         SQLiteDatabase db = getReadableDatabase();
-        Cursor c = db.rawQuery("SELECT * FROM Alunos;", null);
+        Cursor c = db.rawQuery("SELECT * FROM Alunos WHERE desativado = 0;", null);
 
         List<Aluno> alunos = populaAlunos(c);
 
@@ -129,6 +138,7 @@ public class AlunoDAO extends SQLiteOpenHelper{
         dados.put("nota", aluno.getNota());
         dados.put("caminhoFoto", aluno.getCaminhoFoto());
         dados.put("sincronizado", aluno.getSincronizado());
+        dados.put("desativado", aluno.getDesativado());
         return dados;
     }
 
@@ -147,6 +157,7 @@ public class AlunoDAO extends SQLiteOpenHelper{
             aluno.setNota(c.getDouble(c.getColumnIndex("nota")));
             aluno.setCaminhoFoto(c.getString(c.getColumnIndex("caminhoFoto")));
             aluno.setSincronizado(c.getInt(c.getColumnIndex("sincronizado")));
+            aluno.setDesativado(c.getInt(c.getColumnIndex("desativado")));
 
             alunos.add(aluno);
         }
